more region parser into separate file
